+++ src/routes/index.js
@@ -53,6 +53,8 @@ function isStaticRoute(routePattern) {
  * @param {Object} fulfillmentController - The fulfillment controller.
  * @param {Object} importController - The import controller.
  * @param {Object} assistantController - The assistant controller.
+ * @param {Object} apiKeyController - The API key controller.
+ * @param {Object} sitesAuditsController - The sites audits controller.
  * @return {{staticRoutes: {}, dynamicRoutes: {}}} - An object with static and dynamic routes.
  */
 export default function getRouteHandlers(
@@ -67,6 +69,8 @@ export default function getRouteHandlers(
   fulfillmentController,
   importController,
   assistantController,
+  apiKeyController,
+  sitesAuditsToggleController,
 ) {
   const staticRoutes = {};
   const dynamicRoutes = {};
@@ -77,6 +81,7 @@ export default function getRouteHandlers(
     'GET /configurations/latest': configurationController.getLatest,
     'PUT /configurations/latest': configurationController.updateConfiguration,
     'GET /configurations/:version': configurationController.getByVersion,
+    'PATCH /configurations/sites/audits': sitesAuditsToggleController.execute,
     'POST /event/fulfillment': fulfillmentController.processFulfillmentEvents,
     'POST /hooks/site-detection/cdn/:hookSecret': hooksController.processCDNHook,
     'POST /hooks/site-detection/rum/:hookSecret': hooksController.processRUMHook,
@@ -113,6 +118,9 @@ export default function getRouteHandlers(
     'POST /slack/events': slackController.handleEvent,
     'POST /slack/channels/invite-by-user-id': slackController.inviteUserToChannel,
     'GET /trigger': triggerHandler,
+    'POST /tools/api-keys': apiKeyController.createApiKey,
+    'DELETE /tools/api-keys/:id': apiKeyController.deleteApiKey,
+    'GET /tools/api-keys': apiKeyController.getApiKeys,
     'POST /tools/import/jobs': importController.createImportJob,
     'GET /tools/import/jobs/:jobId': importController.getImportJobStatus,
     'DELETE /tools/import/jobs/:jobId': importController.deleteImportJob,